+++ webhook/queue/worker.go
@@ -1,10 +1,15 @@
 package queue
 
+/*
+* This package is used for queuing the webhooks to send using golang channels. In this package,
+we also handle the retries and the exponential backoff logic for sending the hooks.
+*/
+
 import (
 	"context"
 	"fmt"
-	"log"
 	"time"
+	"webhook/logging"
 	"webhook/sender"
 
 	redisClient "webhook/redis"
@@ -27,8 +32,19 @@ func ProcessWebhooks(ctx context.Context, webhookQueue chan redisClient.WebhookP
 
 func sendWebhookWithRetries(payload redisClient.WebhookPayload) {
 	if err := retryWithExponentialBackoff(payload); err != nil {
-		log.Println("Failed to send webhook after maximum retries. WebhookID:", payload.WebhookId)
+		logging.WebhookLogger(logging.WarningType, fmt.Errorf("failed to send webhook after maximum retries. WebhookID : %s", payload.WebhookId))
 	}
+}
+
+func calculateBackoff(currentBackoff time.Duration) time.Duration {
+
+	nextBackoff := currentBackoff * 2
+
+	if nextBackoff > maxBackoff {
+		return maxBackoff
+	}
+
+	return nextBackoff
 }
 
 func retryWithExponentialBackoff(payload redisClient.WebhookPayload) error {
@@ -36,9 +52,9 @@ func retryWithExponentialBackoff(payload redisClient.WebhookPayload) error {
 	backoffTime := initialBackoff
 
 	for retries < maxRetries {
-		err := sender.SendWebhook(payload.Data, payload.Url, payload.WebhookId)
+		err := sender.SendWebhook(payload.Data, payload.Url, payload.WebhookId, payload.SecretHash)
 
-		fmt.Errorf("Error sending webhook:", err)
+		logging.WebhookLogger(logging.ErrorType, fmt.Errorf("error sending webhook: %s", err))
 
 		backoffTime = calculateBackoff(backoffTime)
 		retries++
@@ -46,8 +62,9 @@ func retryWithExponentialBackoff(payload redisClient.WebhookPayload) error {
 		time.Sleep(backoffTime)
 	}
 
-	fmt.Errorf("Maximum retries reached:", maxRetries)
+	logging.WebhookLogger(logging.WarningType, fmt.Errorf("maximum retries reached: %s", maxRetries))
 
+	return nil
 }
 
 func calculateBackoff(currentBackoff time.Duration) time.Duration {