+++ CHANGELOG.md
@@ -9,4 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Refactor ProcessWebhooks for clarity, safety, and testability (#3 )
+- Implement Payload Signing with webhook-signature Header in SendWebhook (#6)
+- Implement File Logging Module for Enhanced Error Handling (#7)
+- Refactor Subscribe function in the redis package for better error handling and modularity( #4 )
 - Improve Payload Flexibility to Accept Various JSON Formats( #1 )